+++ app/Http/Controllers/HomeController.php
@@ -6,8 +6,22 @@ use Illuminate\Http\Request;
 
 class HomeController extends Controller
 {
+    const blade_view = 'layout.home';
+    const url_redirect ='home';
+    const name = 'home';
+
     public function index()
     {
-        return view('layout.home.main');
+        $config = [
+            'blade'     => self::blade_view,
+            'url'       => self::url_redirect
+        ];
+
+        // penguraian data
+        $params = [
+            'config'    => $config
+        ];
+
+        return view('layout.home.main', $params);
     }
 }