+++ resources/views/component/navbar/menu.blade.php
@@ -15,6 +15,7 @@
                         <ul class="dropdown-menu">
                             <li><a href="{{ url('branch') }}" class="dropdown-item">@svg('heroicon-s-building-office-2', 'icon-sm me-1') Branch</a></li>
                             <li><a href="{{ url('branch/coordinator') }}" class="dropdown-item">@svg('heroicon-s-user-group', 'icon-sm me-1') Coordinator</a></li>
+                            <li><a href="{{ url('branch/service') }}" class="dropdown-item">@svg('heroicon-s-building-storefront', 'icon-sm me-1') Service</a></li>
                         </ul>
                     </li>
                 @endif