+++ resources/views/layout/branch/sidebar.blade.php
@@ -1,4 +1,4 @@
-<div class="bg-white">
+<div class="main-sidebar">
     @include('sidebar.profile')
     <ul class="nav nav-pills flex-column" id="sidebar">
         <li class="nav-item mb-1">
@@ -6,6 +6,12 @@
         </li>
         <li class="nav-item mb-1">
             <a href="{{ $type == 'create' ? '#main' : url('branch/create') }}" class="nav-link">@svg('heroicon-o-plus', 'icon') Create new</a>
+        </li>
+        <li class="nav-item mb-1">
+            <a href="{{ url('branch/coordinator') }}" class="nav-link">@svg('heroicon-o-user-group', 'icon') Coordinator</a>
+        </li>
+        <li class="nav-item mb-1">
+            <a href="{{ url('branch/service') }}" class="nav-link">@svg('heroicon-o-building-storefront', 'icon') Service</a>
         </li>
         {{--            Jika all access maka munculkan element--}}
         @if(getUserLevel('branches') >= ALL_ACCESS)