+++ resources/views/sidebar/profile.blade.php
@@ -17,3 +17,4 @@
         </ul>
     </div>
 </div>
+<hr class="border-dashed">